refactor(admin-ui): extract tuning types to types/tuning.ts (behavior-preserving)
TuningRuleModal.tsx(916行) から型定義4種を types/tuning.ts に抽出。
re-export で既存 import パス互換を維持（呼び出し側変更なし）。
Phase2-P1-2 巨大ファイルリファクタ 第1段階。
