+++ admin-ui/src/components/tuning/TuningRuleModal.tsx
@@ -2,31 +2,15 @@ import { useState, useEffect } from "react";
 import { useLang } from "../../i18n/LangContext";
 import { authFetch, API_BASE } from "../../lib/api";
 
-export interface ApprovedResponse {
-  text: string;
-  style: string;
-  reason?: string;
-  approved_at: string;
-}
+import type {
+  ApprovedResponse,
+  TuningRule,
+  TuningRuleInput,
+  SourceConversation,
+} from "../../types/tuning";
 
-export interface TuningRule {
-  id: number;
-  tenant_id: string;
-  trigger_pattern: string;
-  expected_behavior: string;
-  priority: number;
-  is_active: boolean;
-  created_by: string;
-  created_at: string;
-  approved_responses?: ApprovedResponse[];
-}
-
-export type TuningRuleInput = Omit<TuningRule, "id" | "created_by" | "created_at">;
-
-export interface SourceConversation {
-  userMsg: string;
-  assistantMsg: string;
-}
+// 既存の import パス互換のため re-export（呼び出し側は変更不要）
+export type { ApprovedResponse, TuningRule, TuningRuleInput, SourceConversation };
 
 interface TestResponseItem {
   style: string;
